# web/src/lib/openapi.schema.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ScanCompleteEvent": {
        "properties": {
          "success": {
            "title": "Success",
            "type": "boolean"
          },
          "summary": {
            "additionalProperties": true,
            "title": "Summary",
            "type": "object"
          },
          "type": {
            "const": "complete",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "type",
          "success",
          "summary"
        ],
        "title": "ScanCompleteEvent",
        "type": "object"
      },
      "ScanErrorEvent": {
        "properties": {
          "message": {
            "title": "Message",
            "type": "string"
          },
          "phase": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ScanPhase"
              },
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Phase"
          },
          "type": {
            "const": "error",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "type",
          "message"
        ],
        "title": "ScanErrorEvent",
        "type": "object"
      },
      "ScanEvent": {
        "anyOf": [
          {
            "$ref": "#/components/schemas/ScanStatusEvent"
          },
          {
            "$ref": "#/components/schemas/ScanFilesFoundEvent"
          },
          {
            "$ref": "#/components/schemas/ScanProgressEvent"
          },
          {
            "$ref": "#/components/schemas/ScanQueryResultEvent"
          },
          {
            "$ref": "#/components/schemas/ScanRegistryEvent"
          },
          {
            "$ref": "#/components/schemas/ScanCompleteEvent"
          },
          {
            "$ref": "#/components/schemas/ScanErrorEvent"
          }
        ]
      },
      "ScanFilesFoundEvent": {
        "properties": {
          "files": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Files",
            "type": "array"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          },
          "type": {
            "const": "files_found",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "type",
          "files",
          "total"
        ],
        "title": "ScanFilesFoundEvent",
        "type": "object"
      },
      "ScanJsonError": {
        "properties": {
          "error": {
            "title": "Error",
            "type": "string"
          },
          "success": {
            "title": "Success",
            "type": "boolean"
          }
        },
        "required": [
          "success",
          "error"
        ],
        "title": "ScanJsonError",
        "type": "object"
      },
      "ScanJsonSuccess": {
        "properties": {
          "files": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Files",
            "type": "array"
          },
          "queries": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Queries",
            "type": "array"
          },
          "success": {
            "title": "Success",
            "type": "boolean"
          },
          "summary": {
            "additionalProperties": true,
            "title": "Summary",
            "type": "object"
          }
        },
        "required": [
          "success",
          "files",
          "queries",
          "summary"
        ],
        "title": "ScanJsonSuccess",
        "type": "object"
      },
      "ScanPhase": {
        "enum": [
          "config",
          "discovery",
          "extraction",
          "conversion",
          "registry",
          "analysis"
        ],
        "title": "ScanPhase",
        "type": "string"
      },
      "ScanProgressEvent": {
        "properties": {
          "current": {
            "title": "Current",
            "type": "integer"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "phase": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ScanPhase"
              },
              {
                "type": "string"
              }
            ],
            "title": "Phase"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          },
          "type": {
            "const": "progress",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "type",
          "phase",
          "current",
          "total",
          "message"
        ],
        "title": "ScanProgressEvent",
        "type": "object"
      },
      "ScanQueryResultEvent": {
        "properties": {
          "query": {
            "additionalProperties": true,
            "title": "Query",
            "type": "object"
          },
          "type": {
            "const": "query_result",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "type",
          "query"
        ],
        "title": "ScanQueryResultEvent",
        "type": "object"
      },
      "ScanRegistryEvent": {
        "properties": {
          "new_queries": {
            "title": "New Queries",
            "type": "integer"
          },
          "registry_path": {
            "default": "",
            "title": "Registry Path",
            "type": "string"
          },
          "skipped": {
            "title": "Skipped",
            "type": "boolean"
          },
          "total_queries": {
            "title": "Total Queries",
            "type": "integer"
          },
          "type": {
            "const": "registry",
            "title": "Type",
            "type": "string"
          },
          "updated_queries": {
            "title": "Updated Queries",
            "type": "integer"
          }
        },
        "required": [
          "type",
          "new_queries",
          "updated_queries",
          "total_queries",
          "skipped"
        ],
        "title": "ScanRegistryEvent",
        "type": "object"
      },
      "ScanRequest": {
        "description": "Request body for scan endpoint.",
        "properties": {
          "analyze": {
            "default": false,
            "title": "Analyze",
            "type": "boolean"
          },
          "check": {
            "default": false,
            "title": "Check",
            "type": "boolean"
          },
          "diff": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Diff"
          },
          "directory": {
            "title": "Directory",
            "type": "string"
          },
          "dry_run": {
            "default": false,
            "title": "Dry Run",
            "type": "boolean"
          },
          "fail_threshold": {
            "default": 40,
            "title": "Fail Threshold",
            "type": "integer"
          },
          "file_pattern": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "File Pattern"
          },
          "nosave": {
            "default": false,
            "title": "Nosave",
            "type": "boolean"
          },
          "shallow": {
            "default": false,
            "title": "Shallow",
            "type": "boolean"
          },
          "target": {
            "title": "Target",
            "type": "string"
          },
          "warn_threshold": {
            "default": 60,
            "title": "Warn Threshold",
            "type": "integer"
          }
        },
        "required": [
          "target",
          "directory"
        ],
        "title": "ScanRequest",
        "type": "object"
      },
      "ScanStatusEvent": {
        "properties": {
          "message": {
            "title": "Message",
            "type": "string"
          },
          "phase": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ScanPhase"
              },
              {
                "type": "string"
              }
            ],
            "title": "Phase"
          },
          "type": {
            "const": "status",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "type",
          "phase",
          "message"
        ],
        "title": "ScanStatusEvent",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "API server for RDST web client",
    "title": "RDST API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/scan": {
      "post": {
        "description": "Scan a directory for ORM queries via SSE streaming.\n\nStreams real-time progress as the scan progresses through phases:\n- `status`: Phase status messages\n- `files_found`: Files with ORM patterns discovered\n- `progress`: Per-phase progress (current/total)\n- `query_result`: Individual query results\n- `registry`: Registry save results\n- `complete`: Scan complete with summary\n- `error`: Error occurred",
        "operationId": "scan_directory_api_scan_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScanRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              },
              "text/event-stream": {
                "schema": {
                  "$ref": "#/components/schemas/ScanEvent"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Scan Directory",
        "tags": [
          "scan"
        ]
      }
    },
    "/api/scan/json": {
      "post": {
        "description": "Scan a directory and return JSON response (non-streaming).\n\nCollects all events and returns the final summary.",
        "operationId": "scan_directory_json_api_scan_json_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScanRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/ScanJsonSuccess"
                    },
                    {
                      "$ref": "#/components/schemas/ScanJsonError"
                    }
                  ],
                  "title": "Response Scan Directory Json Api Scan Json Post"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Scan Directory Json",
        "tags": [
          "scan"
        ]
      }
    },
...
  }
}